bizflow-project/backend/requirements.txt
Flask>=2.0
Flask-Cors>=3.0
Flask-SQLAlchemy>=2.5
SQLAlchemy>=1.4
marshmallow>=3.0
pymssql>=2.2
psycopg2-binary>=2.9
redis>=4.5
python-dotenv>=0.21
Flask-RESTX>=1.1.0
flasgger
fastapi
apispec
apispec_webframeworks
flask-swagger-ui
openai>=1.0
google-generativeai>=0.3
google-cloud-speech>=2.16
whisper-openai>=0.1
bcrypt>=4.0
PyJWT>=2.6
pandas>=1.5
reportlab>=4.0
tiktoken>=0.5
sentence-transformers>=2.2
